+++ uikit/src/components/KAccordion/__tests__/helpers.tsx
@@ -1,0 +1,4 @@
+import * as KAccordionStories from '../__stories__/KAccordion.stories';
+import { composeStories } from '@storybook/react-vite';
+
+export const KAccordionComposedStories = composeStories(KAccordionStories);