+++ uikit/src/components/KButtonList/__tests__/helpers.tsx
@@ -1,4 +1,0 @@
-import * as KButtonListStories from '../__stories__/KButtonList.stories';
-import { composeStories } from '@storybook/react-vite';
-
-export const KButtonListComposedStories = composeStories(KButtonListStories);