+++ uikit/src/components/KButtonList/index.ts
@@ -1,2 +1,0 @@
-export type { KButtonListProps } from './types';
-export { KButtonList } from './KButtonList';