+++ uikit/src/components/KList/__tests__/helpers.tsx
@@ -1,0 +1,4 @@
+import * as KListStories from '../__stories__/KList.stories';
+import { composeStories } from '@storybook/react-vite';
+
+export const KListComposedStories = composeStories(KListStories);