+++ uikit/src/components/KList/index.ts
@@ -1,0 +1,2 @@
+export type { KListProps } from './types';
+export { KList } from './KList';