+++ uikit/src/components/KTitle/types.ts
@@ -1,9 +1,6 @@
-import type { KIconProps } from '../KIcon';
+import type { KButtonProps } from '../KButton';
 
-export type KTitleAction = {
-  icon?: KIconProps['name'];
-  onClick?: () => void;
-};
+export type KTitleAction = Pick<KButtonProps, 'iconLeft' | 'onClick'>;
 
 export type KTitleProps = {
   children?: React.ReactNode;